chore: shift routes related to users to user.routes.ts
